% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpgf.R
\name{get_cpgf}
\alias{get_cpgf}
\title{Download da CPGF}
\usage{
get_cpgf(year, month)
}
\arguments{
\item{month}{Valid options are 01 a 12.}

\item{year.}{Valid options are 2013 to corrent.}
}
\value{
\code{(get_cpgf)} returns a \code{data.frame}
}
\description{
Download da CPGF
}
\examples{
\dontrun{cartao <- get_cpgf(2023, 12)}

}
